double => float

+++ lib/TempMon/ITempMonAdapter.h
@@ -5,5 +5,5 @@ class ITempMonAdapter
 {
     public:
         virtual ~ITempMonAdapter() {}
-        virtual void NotifyTempChg(double temp) = 0;
+        virtual void NotifyTempChg(float temp) = 0;
 };
